Remove deprecated code
